NF: Remove SyncStatus
The enum was used only in one function. Adding an intermediary step
does not add clarity to the code.

Each method were also only used once.

I think the resulting code is easier to read, thanks to less layer of abstractions.
